# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OrientTogo API",
        "description": "Documentation officielle de l'API OrientTogo (Laravel 13). \n    Toutes les requêtes doivent être préfixées par `/api/v1`. \n    L'authentification se fait via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Équipe Technique OrientTogo",
            "email": "contact@orienttogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "/api",
            "description": "Serveur Local (Base Path)"
        }
    ],
    "paths": {
...
        "/v1/establishments": {
            "get": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Liste des establishments",
                "description": "Retourne une liste paginée avec filtres optionnels",
                "operationId": "getEstablishments",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d’éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EstablishmentResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Créer un(e) Establishment",
                "operationId": "createEstablishment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEstablishmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Establishment créé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EstablishmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/establishments/{idOrSlug}": {
            "get": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Détails d’un(e) Establishment",
                "operationId": "getEstablishment",
                "parameters": [
                    {
                        "name": "idOrSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Establishment trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EstablishmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/establishments/{establishment}": {
            "put": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Mettre à jour un(e) Establishment",
                "operationId": "updateEstablishment",
                "parameters": [
                    {
                        "name": "establishment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEstablishmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Establishment mis(e) à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EstablishmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Supprimer un(e) Establishment",
                "operationId": "deleteEstablishment",
                "parameters": [
                    {
                        "name": "establishment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Establishment supprimé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/establishments/{establishment}/verify": {
            "post": {
                "tags": [
                    "Establishment"
                ],
                "summary": "Valider un établissement (Admin)",
                "operationId": "verifyEstablishment",
                "parameters": [
                    {
                        "name": "establishment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ est obligatoire."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "StoreEstablishmentRequest": {
                "required": [
                    "name",
                    "type",
                    "region",
                    "city",
                    "address",
                    "email",
                    "phone",
                    "website",
                    "description",
                    "fees_info",
                    "specialties",
                    "is_public",
                    "is_verified",
                    "verification_status",
                    "is_selected",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "type": {
                        "type": "string",
                        "example": "Exemple de type"
                    },
                    "region": {
                        "type": "string",
                        "example": "Exemple de region"
                    },
                    "city": {
                        "type": "string",
                        "example": "Exemple de city"
                    },
                    "address": {
                        "type": "string",
                        "example": "Exemple de address"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+22890123456"
                    },
                    "website": {
                        "type": "string",
                        "example": "Exemple de website"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "fees_info": {
                        "type": "string",
                        "example": "Exemple de fees_info"
                    },
                    "specialties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "verification_status": {
                        "type": "string",
                        "example": "Exemple de verification_status"
                    },
                    "is_selected": {
                        "type": "boolean",
                        "example": true
                    },
                    "latitude": {
                        "type": "number",
                        "example": 99.99
                    },
                    "longitude": {
                        "type": "number",
                        "example": 99.99
                    }
                },
                "type": "object"
            },
...
            "UpdateEstablishmentRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "type": {
                        "type": "string",
                        "example": "Exemple de type"
                    },
                    "region": {
                        "type": "string",
                        "example": "Exemple de region"
                    },
                    "city": {
                        "type": "string",
                        "example": "Exemple de city"
                    },
                    "address": {
                        "type": "string",
                        "example": "Exemple de address"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+22890123456"
                    },
                    "website": {
                        "type": "string",
                        "example": "Exemple de website"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "fees_info": {
                        "type": "string",
                        "example": "Exemple de fees"
                    },
                    "specialties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Maths",
                            "Physique"
                        ]
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "verification_status": {
                        "type": "string",
                        "example": "Exemple de verification_status"
                    },
                    "is_selected": {
                        "type": "boolean",
                        "example": true
                    },
                    "latitude": {
                        "type": "number",
                        "example": 99.99
                    },
                    "longitude": {
                        "type": "number",
                        "example": 99.99
                    }
                },
                "type": "object"
            },
...
            "EstablishmentResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Exemple de name"
                    },
                    "type": {
                        "type": "string",
                        "example": "Exemple de type"
                    },
                    "region": {
                        "type": "string",
                        "example": "Exemple de region"
                    },
                    "city": {
                        "type": "string",
                        "example": "Exemple de city"
                    },
                    "address": {
                        "type": "string",
                        "example": "Exemple de address"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+22890123456"
                    },
                    "website": {
                        "type": "string",
                        "example": "Exemple de website"
                    },
                    "description": {
                        "type": "string",
                        "example": "Exemple de description"
                    },
                    "fees": {
                        "type": "string",
                        "example": "Exemple de fees"
                    },
                    "specialties": {
                        "type": "string",
                        "example": "Exemple de specialties"
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "verification_status": {
                        "type": "string",
                        "example": "Exemple de verification_status"
                    },
                    "est_selectionne": {
                        "type": "boolean",
                        "example": true
                    },
                    "latitude": {
                        "type": "number",
                        "example": 99.99
                    },
                    "longitude": {
                        "type": "number",
                        "example": 99.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utilisez le token retourné lors de la connexion (login ou verify OTP).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}